refactor of activeGrid
