+++ src/components/World/ActiveObjPlayerGrid.jsx
@@ -1,22 +1,45 @@
 /* eslint-disable jsx-a11y/click-events-have-key-events, react/prop-types */
-import React, {
-  useState,
-  useCallback,
-  useEffect,
-  useContext,
-  useRef,
-} from 'react';
-import {
-  numCols,
-  numRows,
-  genGridArray,
-  rowFromIndex,
-  colFromIndex,
-} from './GridConstants.mjs';
+import React, { useState, useEffect, useContext, useRef } from 'react';
+import { numCols, numRows, genGridArray } from './GridConstants.mjs';
 import { SocketContext } from '../../contexts/sockets.mjs';
 import { getRandomInt, getUserIdCookie } from '../../../utils.mjs';
-import UserModal from './UserModal.jsx';
-import { updateWorldInDb } from './axiosRequests.jsx';
+import GridSquares from './GridSquares.jsx';
+
+// const directionalKeyPresses = ['KeyA', 'KeyW', 'KeyD', 'KeyS'];
+
+const directionalKeys = {
+  KeyW: {
+    vector: [0, -1],
+    direction: 'top',
+  },
+  KeyS: {
+    vector: [0, 1],
+    direction: 'bottom',
+  },
+  KeyA: {
+    vector: [-1, 0],
+    direction: 'left',
+  },
+  KeyD: {
+    vector: [1, 0],
+    direction: 'right',
+  },
+};
+
+const getDirVectorFromKeyCode = (code) => directionalKeys[code].vector;
+const getDirectionVectors = () => {
+  const keys = Object.keys(directionalKeys);
+  return keys.map((key) => getDirVectorFromKeyCode(key));
+};
+const directions = getDirectionVectors();
+
+const getGridObject = (x, y, grid) => grid[y][x];
+const isWall = (obj) => {
+  if (obj !== null && obj.color !== '') {
+    return true;
+  }
+  return false;
+};
 
 const movedPosition = (userPosition, x, y) => {
   let destinationX = userPosition.x;
@@ -32,340 +55,6 @@ const movedPosition = (userPosition, x, y) => {
   }
 
   return null;
-};
-const directionalKeyPresses = ['KeyA', 'KeyW', 'KeyD', 'KeyS'];
-
-const directions = {
-  top: [0, -1],
-  bottom: [0, 1],
-  left: [-1, 0],
-  right: [1, 0],
-};
-
-const getDisplacementFromKey = (code) => {
-  let direction = [];
-  switch (code) {
-    case 'KeyA':
-      direction = directions.left;
-      break;
-    case 'KeyW':
-      direction = directions.top;
-      break;
-    case 'KeyD':
-      direction = directions.right;
-      break;
-    case 'KeyS':
-      direction = directions.bottom;
-      break;
-    default:
-      direction = [];
-      break;
-  }
-  return direction;
-};
-
-const getGridObject = (x, y, grid) => grid[y][x];
-const isWall = (obj) => {
-  if (obj !== null && obj.color !== '') {
-    return true;
-  }
-  return false;
-};
-
-const movePlayer = (x, y, oldX, oldY, backgrndArr, setUserPosition) => {
-  const gridObj = getGridObject(x, y, backgrndArr);
-  // check if player is hitting any walls
-
-  if (!isWall(gridObj)) {
-    setUserPosition({ x, y });
-  } else {
-    // if user is in the wall
-    const currentGrid = getGridObject(oldX, oldY, backgrndArr);
-    if (isWall(currentGrid)) {
-      setUserPosition({ x, y });
-    }
-  }
-};
-const getAdjCells = (userPosition) => {
-  const directionValues = Object.values(directions);
-  return directionValues.map((dir) =>
-    movedPosition(userPosition, dir[0], dir[1])
-  );
-};
-
-const getActiveAdjCells = (userPosition, activeCells) => {
-  const adjCells = getAdjCells(userPosition).filter((x) => x !== null);
-  const activeAdjCells = [];
-  for (let i = 0; i < activeCells.length; i += 1) {
-    const activeCell = activeCells[i];
-
-    for (let j = 0; j < adjCells.length; j += 1) {
-      const adjCell = adjCells[j];
-      if (activeCell.x === adjCell[0] && activeCell.y === adjCell[1]) {
-        activeAdjCells.push(activeCell);
-      }
-    }
-  }
-  return activeAdjCells;
-};
-const getAdjacentPlayers = (userPosition, playersPositions) => {
-  const adjCells = getAdjCells(userPosition).filter((x) => x !== null);
-  const adjPlayers = [];
-  for (let j = 0; j < adjCells.length; j += 1) {
-    const adjCell = adjCells[j];
-    const [x, y] = adjCell;
-    const playerCell = playersPositions[y][x];
-
-    if (playerCell !== null) {
-      adjPlayers.push(playerCell);
-    }
-  }
-
-  return adjPlayers;
-};
-const clickOnCell = (obj) => {
-  window.open(obj.url);
-};
-const iconFromObjType = (obj) => {
-  let img = '';
-  switch (obj.type) {
-    case 'zoom':
-      img = 'zoom.png';
-      break;
-    case 'note':
-      img = 'sticky_notes.png';
-      break;
-    case 'lucid':
-      img = 'lucid.png';
-      break;
-    case 'figma':
-      img = 'figma.png';
-      break;
-    default:
-      break;
-  }
-
-  return img;
-};
-const clickOnPlayer = (player) => {
-  console.log(`This is player ${player}`);
-};
-
-const buildOnCell = (index, world, setWorld, buildTool, socket) => {
-  const row = rowFromIndex(index);
-  const col = colFromIndex(index);
-  const { board } = world.worldState;
-  if (buildTool.tool === 'wall') {
-    world.worldState.board[row][col] = {
-      ...board[row][col],
-      color: buildTool.color,
-    };
-    world.worldState.board = board;
-    // why not working?
-
-    console.log('world.worldStates going into db :>> ', world);
-    updateWorldInDb(world.id, world.worldState);
-    socket.emit('grid:update:world', { worldId: world.id });
-
-    setWorld({ ...world });
-  }
-  // build tool
-  // {
-  //   tool: '',
-  //   color: '',
-  //   roomId: 0,
-  //   charFill: '',
-  //   activeObjType: '',
-  //   url: '',
-  // }
-};
-
-const Square = ({
-  player,
-  index,
-  userId,
-  userSquare,
-  world,
-  setWorld,
-  buildTool,
-  setBuildTool,
-  socket,
-}) => {
-  const buildToolType = buildTool.type;
-  // need to get x and y coordinate to update world board
-  let fill = (
-    // eslint-disable-next-line jsx-a11y/control-has-associated-label
-    // eslint-disable-next-line jsx-a11y/no-static-element-interactions
-    <div
-      className="cell gridBorder"
-      onClick={() => {
-        if (buildToolType !== '') {
-          buildOnCell(index, world, setWorld, buildTool, socket);
-        }
-      }}
-    />
-  );
-  if (player !== null) {
-    if (typeof player === 'object') {
-      fill = (
-        <input
-          className="cell gridBorder"
-          type="image"
-          src={iconFromObjType(player)}
-          onClick={() => {
-            if (buildToolType === '') {
-              clickOnCell(player);
-            } else {
-              buildOnCell(index, world, setWorld, buildTool, socket);
-            }
-          }}
-          key={`active${index}`}
-          alt={player.type}
-        />
-      );
-    } else if (player === userId) {
-      fill = (
-        // eslint-disable-next-line jsx-a11y/no-static-element-interactions
-        <div
-          onClick={() => {
-            if (buildToolType === '') {
-              clickOnPlayer(player);
-            }
-          }}
-          ref={userSquare}
-          type="image"
-          className="cell gridBorder"
-          style={{
-            backgroundImage: `url("https://avatars.dicebear.com/api/big-smile/${
-              player + 1
-            }.svg")`,
-            cursor: 'pointer',
-            position: 'relative',
-          }}
-        >
-          <UserModal userSquare={userSquare} setBuildTool={setBuildTool} />
-        </div>
-      );
-    } else {
-      fill = (
-        <input
-          onClick={() => {
-            if (buildToolType === '') {
-              clickOnPlayer(player);
-            } else {
-              buildOnCell(index, world, setWorld, buildTool, socket);
-            }
-          }}
-          type="image"
-          className="cell gridBorder"
-          src={`https://avatars.dicebear.com/api/big-smile/${player + 1}.svg`}
-          alt="profile pic"
-        />
-      );
-    }
-  }
-
-  return fill;
-};
-
-const GridSquares = ({
-  activeCells,
-  playersPositions,
-  userSquare,
-  userId,
-  world,
-  setWorld,
-  buildTool,
-  setBuildTool,
-  socket,
-}) => {
-  activeCells.forEach((activity) => {
-    const activityObj = {
-      type: activity.type,
-      url: activity.url,
-    };
-    playersPositions[activity.y][activity.x] = {
-      ...activityObj,
-    };
-  });
-
-  const playerPos1d = [].concat(...playersPositions);
-  const squares = playerPos1d.map((player, i) => (
-    <Square
-      key={`sq_${i.toString()}`}
-      index={i}
-      player={player}
-      userId={userId}
-      userSquare={userSquare}
-      world={world}
-      setWorld={setWorld}
-      buildTool={buildTool}
-      setBuildTool={setBuildTool}
-      socket={socket}
-    />
-  ));
-
-  return (
-    <div
-      id="baseGrid"
-      className="grid-container position-absolute position-absolute-stretch"
-    >
-      {squares}
-    </div>
-  );
-};
-
-const handleDirKeys = (key, userPosition, backgrndArr, setUserPosition) => {
-  const direction = getDisplacementFromKey(key);
-
-  const movedLocation = movedPosition(userPosition, direction[0], direction[1]);
-  if (movedLocation !== null) {
-    movePlayer(
-      movedLocation[0],
-      movedLocation[1],
-      userPosition.x,
-      userPosition.y,
-      backgrndArr,
-      setUserPosition
-    );
-  }
-};
-const openInteractiveObj = (userPosition, activeCells) => {
-  const activeAdjCells = getActiveAdjCells(userPosition, activeCells);
-
-  activeAdjCells.forEach((cell) => {
-    window.open(cell.url);
-  });
-};
-const interactWPlayer = (userPosition, playersPositions, userId) => {
-  let adjPlayerCell = getAdjacentPlayers(userPosition, playersPositions);
-  adjPlayerCell = adjPlayerCell.filter((player) => player !== userId);
-
-  adjPlayerCell.forEach((player) => {
-    console.log(`adjacent to player ${player}`);
-  });
-};
-const handleInteractKey = (
-  userPosition,
-  activeCells,
-  playersPositions,
-  userId
-) => {
-  openInteractiveObj(userPosition, activeCells);
-  interactWPlayer(userPosition, playersPositions, userId);
-};
-const handleBuildKey = (userPosition, userSquare, buildTool, setBuildTool) => {
-  const userSquareChilds = userSquare.current.childNodes;
-  userSquareChilds.forEach((child) => {
-    const currDisplay = child.style.display;
-    child.style.display = currDisplay === 'none' ? 'block' : 'none';
-  });
-
-  const modal = (
-    <div>
-      User at x :{userPosition.x}, y: {userPosition.y}
-    </div>
-  );
 };
 export default function ActiveObjPlayerGrid({
   isChatFocused,
@@ -395,39 +84,137 @@ export default function ActiveObjPlayerGrid({
   const socket = useContext(SocketContext);
   const userId = getUserIdCookie();
   // link with pages/db
+  // ======================START OF FUNCTIONS===============================
 
-  const handlePlayersPositions = useCallback((playerPos) => {
+  const movePlayer = (x, y, oldX, oldY) => {
+    const gridObj = getGridObject(x, y, backgrndArr);
+    // check if player is hitting any walls
+    if (!isWall(gridObj)) {
+      setUserPosition({ x, y });
+    } else {
+      // if user is in the wall
+      const currentGridObj = getGridObject(oldX, oldY, backgrndArr);
+      if (isWall(currentGridObj)) {
+        setUserPosition({ x, y });
+      }
+    }
+  };
+  const getAdjCells = () =>
+    directions.map((dir) => movedPosition(userPosition, dir[0], dir[1]));
+
+  const getAdjacentPlayers = () => {
+    const adjCells = getAdjCells().filter((x) => x !== null);
+    const adjPlayers = [];
+    for (let j = 0; j < adjCells.length; j += 1) {
+      const adjCell = adjCells[j];
+      const [x, y] = adjCell;
+      const playerCell = playersPositions[y][x];
+
+      if (playerCell !== null) {
+        adjPlayers.push(playerCell);
+      }
+    }
+    return adjPlayers;
+  };
+
+  const handleDirKeys = (key) => {
+    const direction = getDirVectorFromKeyCode(key);
+
+    const movedLocation = movedPosition(
+      userPosition,
+      direction[0],
+      direction[1]
+    );
+    if (movedLocation !== null) {
+      movePlayer(
+        movedLocation[0],
+        movedLocation[1],
+        userPosition.x,
+        userPosition.y
+      );
+    }
+  };
+
+  const interactWPlayer = () => {
+    let adjPlayerCell = getAdjacentPlayers();
+    adjPlayerCell = adjPlayerCell.filter((player) => player !== userId);
+
+    adjPlayerCell.forEach((player) => {
+      console.log(`adjacent to player ${player}`);
+    });
+  };
+
+  const getActiveAdjCells = () => {
+    const adjCells = getAdjCells().filter((x) => x !== null);
+    const activeAdjCells = [];
+    for (let i = 0; i < activeCells.length; i += 1) {
+      const activeCell = activeCells[i];
+
+      for (let j = 0; j < adjCells.length; j += 1) {
+        const adjCell = adjCells[j];
+        if (activeCell.x === adjCell[0] && activeCell.y === adjCell[1]) {
+          activeAdjCells.push(activeCell);
+        }
+      }
+    }
+    return activeAdjCells;
+  };
+  const openInteractiveObj = () => {
+    const activeAdjCells = getActiveAdjCells();
+
+    activeAdjCells.forEach((cell) => {
+      window.open(cell.url);
+    });
+  };
+  const handleInteractKey = () => {
+    openInteractiveObj();
+    interactWPlayer();
+  };
+  const handleBuildKey = () => {
+    const userSquareChilds = userSquare.current.childNodes;
+    userSquareChilds.forEach((child) => {
+      const currDisplay = child.style.display;
+      child.style.display = currDisplay === 'none' ? 'block' : 'none';
+    });
+  };
+
+  const handlePlayersPositions = (playerPos) => {
+    console.log('hanndling playing positions');
+    // console.log('playerPos :>> ', playerPos);
+    // remove existing copy
+    playerPos = playerPos.map((row) =>
+      row.map((cell) => (cell === userId ? null : cell))
+    );
     const { x, y } = userPosition;
     playerPos[y][x] = userId;
     setPlayersPositions(playerPos);
-  });
-
+  };
+  // ======================END OF FUNCTIONS===============================
   useEffect(() => {
     // SOCKETS
-    socket.emit('grid:join', { userId, worldId });
+    console.log('userPosition :>> ', userPosition);
+    socket.emit('grid:join', { userId, worldId, userPosition });
   }, [socket]);
 
   useEffect(() => {
+    console.log('in use effect key press handler');
     // KEY PRESSES
     const handleKeyPress = (e) => {
       if (!isChatFocused) {
-        if (directionalKeyPresses.includes(e.code)) {
-          handleDirKeys(e.code, userPosition, backgrndArr, setUserPosition);
-          setBuildTool({
-            tool: '',
-            color: '',
-            roomId: 0,
-            charFill: '',
-            activeObjType: '',
-            url: '',
-          });
+        if (Object.keys(directionalKeys).includes(e.code)) {
+          handleDirKeys(e.code);
+          if (buildTool.tool !== '') {
+            setBuildTool({
+              tool: '',
+              color: '',
+              roomId: 0,
+              charFill: '',
+              activeObjType: '',
+              url: '',
+            });
+          }
         } else if (e.code === 'KeyE') {
-          handleInteractKey(
-            userPosition,
-            activeCells,
-            playersPositions,
-            userId
-          );
+          handleInteractKey();
         } else if (e.code === 'KeyB') {
           // get square of userPosition
           handleBuildKey(userPosition, userSquare, buildTool, setBuildTool);
@@ -435,6 +222,13 @@ export default function ActiveObjPlayerGrid({
       }
     };
     document.addEventListener('keypress', handleKeyPress);
+    return () => {
+      document.removeEventListener('keypress', handleKeyPress);
+    };
+  }, [socket, userPosition, isChatFocused]);
+
+  useEffect(() => {
+    console.log('in use effect emitting and listening');
     // SOCKETS
     socket.emit('grid:update', {
       worldId,
@@ -446,10 +240,9 @@ export default function ActiveObjPlayerGrid({
     socket.on('PLAYER_POSITIONS', handlePlayersPositions);
 
     return () => {
-      document.removeEventListener('keypress', handleKeyPress);
       socket.off('PLAYER_POSITIONS', handlePlayersPositions);
     };
-  }, [socket, userPosition, isChatFocused]);
+  }, [socket, userPosition]);
 
   return (
     <>
@@ -463,7 +256,6 @@ export default function ActiveObjPlayerGrid({
         setWorld={setWorld}
         buildTool={buildTool}
         setBuildTool={setBuildTool}
-        socket={socket}
       />
     </>
   );
